challange finished

+++ src/modules/orders/infra/typeorm/entities/OrdersProducts.ts
@@ -33,7 +33,7 @@ class OrdersProducts {
   @Column({ type: 'decimal', precision: 5, scale: 2 })
   price: number;
 
-  @Column('decimal')
+  @Column('float')
   quantity: number;
 
   @CreateDateColumn()
